+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGE LOG
 
-## [Pre-release]
+## Unreleased
+
+- Fix error when collections are empty
+- Fix error when a collection is empty
+- SFR-2077: Validate top 5 language filters for Frontend DRB search results
+
+## [0.18.2]
 
 - Fix cut off text on search bar dropdown
 - Fix broken link on the About page